requirements.txt
# Requirements for cloud deployment - Python 3.11.7 compatible
# Updated: 2026-01-16 - Removed streamlit-autorefresh for compatibility
streamlit>=1.28.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
plotly>=5.15.0,<6.0.0
ta>=0.10.2,<1.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
asyncpg>=0.29.0
requests>=2.31.0,<3.0.0
yfinance>=0.2.18,<1.0.0
matplotlib>=3.7.0,<4.0.0
openai>=1.0.0,<2.0.0
firebase-admin>=6.2.0,<7.0.0
google-cloud-firestore>=2.11.0,<3.0.0
fastapi>=0.112.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
pydantic>=2.8.0,<3.0.0
joblib>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
mistralai>=0.0.11,<1.0.0
